using time_get_mircoseconds and storing it as int64 val

+++ src/game/client/ui.h
@@ -103,7 +103,7 @@ struct SUIAnimator
 	bool m_ScaleLabel;
 	bool m_RepositionLabel;
 
-	float m_Time;
+	long long m_Time;
 	float m_Value;
 
 	float m_XOffset;
